create individual styled files for each component in the modal_shared folder
