+++ src/components/modal_shared/ButtonContainer.js
@@ -1,4 +1,4 @@
-import * as s from "./Modal.sc";
+import * as s from "./ButtonContainer.sc";
 
 export default function ButtonContainer({ children, buttonCountNum }) {
   return (
