+++ src/components/modal_shared/Footer.js
@@ -1,4 +1,4 @@
-import * as s from "./Modal.sc";
+import * as s from "./Footer.sc";
 
 export default function Footer({ children }) {
   return <s.Footer>{children}</s.Footer>;
