+++ src/components/modal_shared/GoToButton.sc.js
@@ -1,0 +1,21 @@
+import { zeeguuDarkOrange } from "../colors";
+import { OrangeRoundButton } from "../allButtons.sc";
+import styled from "styled-components";
+
+//redesigned button for a better focal point and improved
+//readability of the text inside it.
+//TODO: After implementing all the onboarding steps,
+//create style quide for all buttons and refactor / factor them out
+const GoToButton = styled(OrangeRoundButton)`
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-start;
+  align-items: center;
+  gap: 0.25rem;
+  padding: 0.7em 2em;
+  border-radius: 4em;
+  font-weight: 600;
+  border-bottom: solid 0.2em ${zeeguuDarkOrange};
+`;
+
+export { GoToButton };
