+++ src/components/modal_shared/Header.js
@@ -1,4 +1,4 @@
-import * as s from "./Modal.sc";
+import * as s from "./Header.sc";
 
 export default function Header({ children }) {
   return <s.Header>{children}</s.Header>;
