+++ src/components/modal_shared/Heading.js
@@ -1,4 +1,4 @@
-import * as s from "./Modal.sc";
+import * as s from "./Heading.sc";
 
 export default function Heading({ children }) {
   return <s.Heading>{children}</s.Heading>;
