+++ src/components/modal_shared/Icon.js
@@ -1,4 +1,4 @@
-import * as s from "./Modal.sc";
+import * as s from "./Icon.sc";
 
 export default function Icon({ src, size }) {
   return <s.Icon type={size} src={src} alt=""/>;
