+++ src/components/modal_shared/Icon.sc.js
@@ -1,0 +1,17 @@
+import styled, { css } from "styled-components";
+
+const Icon = styled.img`
+  margin: 0 0.4rem -0.1rem 0;
+  height: 1.2rem;
+  width: 1.2rem;
+
+  ${(props) =>
+    props.type === "small" &&
+    css`
+      margin: 0 0.4rem;
+      height: 1rem;
+      width: 1rem;
+    `}
+`;
+
+export { Icon };
