+++ src/components/modal_shared/MainImage.js
@@ -1,4 +1,4 @@
-import * as s from "./Modal.sc";
+import * as s from "./MainImage.sc";
 
 export default function MainImage({ src, alt }) {
   return <s.MainImage src={src} alt={alt} />;
