+++ src/components/modal_shared/MainImage.sc.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+const MainImage = styled.img`
+  height: 100%;
+  width: 100%;
+  object-fit: contain;
+`;
+
+export { MainImage };
